requirements.txt
Django==4.2
djangorestframework
django-cors-headers
groq
requests
pandas
numpy
pillow
redis
celery
python-dotenv
